requirements.txt
fastapi~=0.68.0
uvicorn==0.15.0
sqlalchemy~=1.4.25
passlib==1.7.4
pyjwt
alembic==1.13.1
pydantic~=1.10.15
psycopg2-binary
python-dotenv~=1.0.1